requirements.txt
SQLAlchemy==1.4.11
Flask==0.12.5
PyYAML==5.4.1
